requirements.txt
aiohttp
aiohttp-retry
asgiref >= 3.5.2, < 4
bs4
celery >= 5.2
django >= 4.1.1
django-celery-beat >= 2.3
django-celery-results >= 2.4
django-dynamic-preferences
django-environ
django-filter
djangorestframework
feedparser
gunicorn
lxml
psycopg2-binary
pyyaml
spacy
uvicorn
